% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{buildFileNameFun}
\alias{buildFileNameFun}
\alias{buildFileName}
\title{Sequentially numbered file name creator}
\usage{
buildFileNameFun(..., prefix = "ponto", ext = ".rda", sep = "-")

buildFileName(n, prefix = "ponto", ext = NULL, sep = "-", nc = NULL)
}
\arguments{
\item{prefix}{Prefix of the filename.}

\item{ext}{File name extension. If \code{NULL}, use no extension.}

\item{sep}{Separator element between different parts of the file name}

\item{n}{Numbers used to build file name.}

\item{nc}{Number of characters to use when building the numbers. If null use largest number in \code{n}.}
}
\value{
Character vector with the names.
}
\description{
Useful functions that create sequentially numbered file
 names.
}
\details{
It is very common useful to build file names with sequential numbers
within. So that the numbers appear in the correct order in directory
listings or simply that when viewing the file names they appear
homogeneous it is nice to have a function that perform build this file
names easily. \code{buildFileName} takes as input a sequence of numbers
and builds the file names with given prefix, extension and
separator.

The other function, \code{buildFileNameFun} returns a
function that creates the file name given the indexes. It allows
multiple indexes so that, for instance a grid of file names can be
created.
}
\examples{
print(buildFileName(1:10, ext='.txt'))
print(buildFileName(1:10, ext='.txt', sep='_'))
print(buildFileName(1:10, ext=NULL))
print(buildFileName(1:10, nc=3))
}
