% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour.R
\name{fidContour}
\alias{fidContour}
\alias{contourf}
\title{Draws filled contour and contour lines.}
\usage{
fidContour(x = seq(0, 1, length.out = nrow(z)), y = seq(0, 1, length.out =
  ncol(z)), z, xlim = range(x, finite = TRUE), ylim = range(y, finite =
  TRUE), zlim = range(z, finite = TRUE), levels = pretty(zlim, nlevels),
  nlevels = 20, color.palette = cm.colors,
  col = color.palette(length(levels) - 1), plot.title, plot.axes, key.title,
  key.axes, asp = NA, xaxs = "i", yaxs = "i", las = 1, axes = TRUE,
  frame.plot = axes, contour.lines = FALSE, nclevels = 10,
  clevels = pretty(zlim, nclevels), labels = NULL, vfont = NULL,
  labcex = 0.6, drawlabels = TRUE, method = "flattest",
  ccol = par("fg"), lty = par("lty"), lwd = par("lwd"), shapeFun = NULL,
  plot.key = TRUE, ...)

contourf(x = seq(0, 1, length.out = nrow(z)), y = seq(0, 1, length.out =
  ncol(z)), z, xlim = range(x, finite = TRUE), ylim = range(y, finite =
  TRUE), zlim = range(z, finite = TRUE), levels = pretty(zlim, nlevels),
  nlevels = 20, color.palette = cm.colors,
  col = color.palette(length(levels) - 1), plot.title, plot.axes, key.title,
  key.axes, asp = NA, xaxs = "i", yaxs = "i", las = 1, axes = TRUE,
  frame.plot = axes, contour.lines = TRUE, nclevels = 10,
  clevels = pretty(zlim, nclevels), labels = NULL, vfont = NULL,
  labcex = 0.6, drawlabels = TRUE, method = "flattest",
  ccol = par("fg"), lty = par("lty"), lwd = par("lwd"), shapeFun = NULL,
  plot.key = TRUE, ...)
}
\arguments{
\item{x, y}{locations of grid lines at which the values in ‘z’ are
measured.  These must be in ascending order.  By default,
equally spaced values from 0 to 1 are used.  If ‘x’ is a
‘list’, its components ‘x$x’ and ‘x$y’ are used for ‘x’ and
‘y’, respectively. If the list has component ‘z’ this is used
for ‘z’.}

\item{z}{A matrix containing the values to be plotted (‘NA’s are
allowed).  Note that ‘x’ can be used instead of ‘z’ for
convenience.}

\item{xlim, ylim, zlim}{x-, y- and z-limits for the plot.}

\item{levels}{a set of levels which are used to partition the range of ‘z’.
Must be *strictly* increasing (and finite).  Areas with ‘z’
values between consecutive levels are painted with the same
color.}

\item{nlevels}{if ‘levels’ is not specified, the range of ‘z’, values is
divided into approximately this many levels.}

\item{color.palette}{a color palette function to be used to assign colors in
the plot.}

\item{col}{an explicit set of colors to be used in the plot.  This
argument overrides any palette function specification.  There
should be one less color than levels}

\item{plot.title}{statements which add titles to the main plot.}

\item{plot.axes}{statements which draw axes (and a ‘box’) on the main plot.
This overrides the default axes.}

\item{key.title}{statements which add titles for the plot key.}

\item{key.axes}{statements which draw axes on the plot key.  This overrides
the default axis.}

\item{asp}{the y/x aspect ratio, see ‘plot.window’.}

\item{xaxs}{the x axis style.  The default is to use internal labeling.}

\item{yaxs}{the y axis style.  The default is to use internal labeling.}

\item{las}{the style of labeling to be used.  The default is to use
horizontal labeling.}

\item{axes, frame.plot}{logicals indicating if axes and a box should be
drawn, as in ‘plot.default’.}

\item{contour.lines}{Plot the contour lines?}

\item{nclevels}{number of contour levels desired *iff* ‘levels’ is not
supplied.}

\item{clevels}{numeric vector of levels at which to draw contour lines.}

\item{labels}{a vector giving the labels for the contour lines.  If ‘NULL’
then the levels are used as labels, otherwise this is coerced
by ‘as.character’.}

\item{vfont}{if ‘NULL’, the current font family and face are used for the
contour labels.  If a character vector of length 2 then
Hershey vector fonts are used for the contour labels. The
first element of the vector selects a typeface and the second
element selects a fontindex (see ‘text’ for more
information).  The default is ‘NULL’ on graphics devices with
high-quality rotation of text and ‘c("sans serif", "plain")’
otherwise.}

\item{labcex}{‘cex’ for contour labelling.  This is an absolute size, not a
multiple of ‘par("cex")’.}

\item{drawlabels}{logical.  Contours are labelled if ‘TRUE’.}

\item{method}{character string specifying where the labels will be located.
Possible values are ‘"simple"’, ‘"edge"’ and ‘"flattest"’
(the default).  See the ‘Details’ section.}

\item{ccol}{color for the lines drawn.}

\item{lty}{line type for the lines drawn.}

\item{lwd}{line width for the lines drawn.}

\item{shapeFun}{Function that plots something on the canvas.}

\item{add}{logical. If ‘TRUE’, add to a current plot.}
}
\description{
Basically a combination of functions \code{\link{contour}} and \code{\link{filled.contour}}.
}
\seealso{
\code{\link{contour}} \code{\link{filled.contour}}
}
