% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.R
\name{cosd}
\alias{cosd}
\alias{sind}
\alias{tand}
\alias{asind}
\alias{acosd}
\alias{atand}
\alias{atan2d}
\title{Trigonometric functions using degrees for angles.}
\usage{
cosd(x)
}
\arguments{
\item{x}{Numeric vectors.}

\item{y}{Numeric vectors.}
}
\value{
Numeric vector.
}
\description{
Implementation of trigonometric functions that uses
degrees instead of radians for angle arguments and
results.
}
\details{
\code{cosd} 
\code{sind} 
\code{tand} 
\code{acosd} 
\code{asind} 
\code{atand}
\code{atan2d}
}
\seealso{
\code{\link{cos}} \code{\link{sin}} \code{\link{tan}} \code{\link{acos}} \code{\link{asin}}
\code{\link{atan}} \code{\link{atan2}}
}
