% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilbert.R
\name{hilbert}
\alias{hilbert}
\title{Hilbert transform
Calculates the Hilbert transform of a vector or columns of a matrix}
\usage{
hilbert(x)
}
\arguments{
\item{x}{Vector or matrix}
}
\value{
Hilbert transform of x
}
\description{
Hilbert transform
Calculates the Hilbert transform of a vector or columns of a matrix
}
\examples{
x <- seq(0, 1, by=0.001)[1:1000]
y <- sin(6*pi*x)
H <- hilbert(x)
plot(y, ty='l', ylim=range(y, Re(H), Im(H)))
lines(abs(H), col=2, lwd=2)
lines(Re(H), col=3)
lines(Im(H), col=4)
}
