% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{integr}
\alias{integr}
\title{Signal integrator and differentiator.}
\usage{
integr(x, p = 1, dt = NULL)
}
\arguments{
\item{x}{Vector or matrix containing data to be integrated. If matrix, each column corresponds to a time series.}

\item{p}{Order of integration. Negative integers represent derivatives.}

\item{dt}{Time steps betweend samples. If not specified, \code{x} is assumed to be a time series \code{\link{ts}} object.}
}
\value{
A vector The integral of the signal.
}
\description{
Implements integrals and derivatives of sampled data using \code{\link{mvfft}}.
}
\details{
Note that the function assumes that the signal has zero mean. The nonzer mean is not integrated.
}
\examples{
tt <- seq(0, 1, len=65)[1:64]
dt <- tt[2]
x <- ts(cos(2*pi*tt), start=0, deltat=dt)
dx <- integr(x, -1)
ix <- integr(x, 1)
plot(x, ylim=range(x, dx, ix))
lines(dx, lty=2)
lines(ix, lty=3)
legend(0, max(x,dx,ix), c("Signal", "Derivative", "Integral"), lty=1:3)
}
