\docType{package}
\name{wutils}
\alias{wutils}
\alias{wutils-package}
\title{Utility functions used by Tom Wallis.}
\description{
  wutils is a utility package of code Tom Wallis commonly
  uses across projects to produce analyses. It includes
  some plotting styles to make plots look nice to Tom, and
  also general helper functions. Not all of the code in
  here was written by Tom; attributions are in source.
}
\details{
  wutils.

  Package: wutils Type: package Version: 0.1 License: GPL-3
}
\author{
  Thomas Wallis \email{thomas.wallis@uni-tuebingen.de}

  including code lifted from the internet, from Winston
  Chang, Simon Barthelme and others.
}

