% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turb.R
\name{linearFit}
\alias{linearFit}
\title{Linear regression.}
\usage{
linearFit(x, y)
}
\arguments{
\item{x}{Vector containing x coordinates of the points.}

\item{y}{Vector containing y coordinates of the points.}
}
\value{
Vector with fit coefficients.
}
\description{
Calculates the linear fit of a set of points using \code{\link{lm}}.
This function returns the coefficients that best approximate the
set of points according to the equation 
\deqn{y = a_1 + a_2\cdot x}{y = a1 + a2*x}
The function returns the vector \code{c(a1, a2)}.
}
\examples{
x <- 1:10
y <- 2*x + rnorm(10, sd=0.4)
fit <- linearFit(x, y)
plot(x, y, xlab='x', ylab='y')
lines(x, fit[1] + fit[2]*x)
print(fit)
}
\seealso{
\code{\link{lm}}  \code{\link{predict.lm}}
}
\author{
Paulo José Saiz Jabardo
}
