% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turb.R
\name{logFit}
\alias{logFit}
\title{Log fit}
\usage{
logFit(x, y)
}
\arguments{
\item{x}{Vector containing x coordinates of the points.}

\item{y}{Vector containing y coordinates of the points.}
}
\value{
Vector with fit coefficients.
}
\description{
Calculates the log fit of a set of points using \code{\link{lm}}.
}
\details{
This function returns the coefficients that best approximate the
set of points according to the equation 
\deqn{y = a_1 + a_2\cdot \log x}{y = a1 + a2*log(x)}
The function returns the vector \code{c(a1, a2)}.
}
\examples{
x <- 1:10
y <- 3*log(x) + 1 + rnorm(10, sd=0.2)

fit <- logFit(x, y)
plot(x, y, xlab='x', ylab='y')
lines(x, fit[1] + fit[2]*log(x))
print(fit)#' x <- 1:10
}
\seealso{
\code{\link{lm}}  \code{\link{predict.lm}}
}
\author{
Paulo José Saiz Jabardo.
}
