\name{rounddx}
\alias{rounddx}
\title{Rounds number to nearest increment.}
\usage{
  rounddx(x, dx = 1)
}
\arguments{
  \item{x}{Number to be rounded.}

  \item{dx}{Increment to be rounded.}
}
\value{
  Rounded number.
}
\description{
  Instead of rounding a number to the nearest nth decimal
  place, this function rounds to the neares multiple of
  \code{dx}.
}

